requirements.txt
# Core ML libraries
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
timm>=0.6.0

# Image processing
opencv-python>=4.5.0
pillow>=8.0.0
numpy==1.26.4

# Configuration
pyyaml>=5.4.0

# Optional: For better performance
# accelerate>=0.12.0  # Uncomment for faster model loadingscipy
